refactor: kafka service logic, infrastructure 분리
